backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magnetar-prometheus"
# Note: This is the PEP-440 semantic version for python packaging.
# The canonical release version stamp for MagnetarPrometheus follows
# the 'yyyy.MM.dd HH:mm:sss' format and is generated separately.
version = "0.1.0"
description = "Backend runtime for the MagnetarPrometheus workflow orchestration platform."
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = [
  "tests",
  "../sdk/python/tests",
]
pythonpath = [
  "src",
  "../sdk/python/src",
]
addopts = "--cov=magnetar_prometheus --cov=magnetar_prometheus_sdk --cov-report=term-missing --cov-fail-under=100"
